some search use case tests
